402: rename sequences for clarity
