+++ web-client/src/presenter/sequences/startACaseConfirmCancelSequence.js
@@ -1,0 +1,5 @@
+import { toggle } from 'cerebral/factories';
+import { state } from 'cerebral';
+import navigateToDashboardAction from '../actions/navigateToDashboardAction';
+
+export default [toggle(state.showModal), navigateToDashboardAction];
