--- web-client/src/presenter/sequences/startACaseConfirmSequence.js
@@ -1,4 +1,0 @@
-import { toggle } from 'cerebral/factories';
-import { state } from 'cerebral';
-
-export default [toggle(state.showModal)];
