+++ web-client/src/presenter/sequences/startACaseToggleCancelSequence.js
